+++ CHANGELOG.md
@@ -19,6 +19,18 @@
 - [](https://github.com/vegaprotocol/vega/issues/xxx)
 
 
+## 0.77.1
+
+### 🛠 Improvements
+
+- [11466](https://github.com/vegaprotocol/vega/issues/11466) - Update CometBFT to version `0.38.10`.
+
+### 🐛 Fixes
+
+- [11453](https://github.com/vegaprotocol/vega/issues/11453) - Fix margin for fully collateralised markets.
+- [11462](https://github.com/vegaprotocol/vega/issues/11462) - Update the time weighted notional when publishing live game scores.
+
+
 ## 0.77.0
 
 ### 🛠 Improvements
@@ -71,7 +83,6 @@
 - [11448](https://github.com/vegaprotocol/vega/issues/11448) - Fix team game score query with epoch filter.
 - [11457](https://github.com/vegaprotocol/vega/issues/11457) - Fix cursor column ordering for game scores.
 - [11454](https://github.com/vegaprotocol/vega/issues/11454) - Ensure ended transfers proper handling.
-
 
 ## 0.76.1